Level population + HUD design
Implemented the different level designs with the overall HUD. Added the functionality to change between weapons and shoot them. Changed the name of playerInformation to shipInformation (this script is just a class and can be used for any agent). The main menu can take you to any level you choose although the character model is not updated as of yet.